fix: fixes DI scope disposing only when message processing ends
