--- src/KafkaFlow.BatchConsume/IWorkerBatch.cs
@@ -1,9 +1,0 @@
-namespace KafkaFlow.BatchConsume
-{
-    using System.Threading.Tasks;
-
-    internal interface IWorkerBatch
-    {
-        Task AddAsync(IMessageContext context, MiddlewareDelegate next);
-    }
-}
