--- src/KafkaFlow.BatchConsume/IWorkerBatchFactory.cs
@@ -1,9 +1,0 @@
-namespace KafkaFlow.BatchConsume
-{
-    using System;
-
-    internal interface IWorkerBatchFactory
-    {
-        IWorkerBatch Create(int batchSize, TimeSpan batchTimeout, ILogHandler logHandler);
-    }
-}
