+++ src/KafkaFlow.UnitTests/BatchConsume/BatchConsumeMiddlewareTests.cs
@@ -2,6 +2,7 @@ namespace KafkaFlow.UnitTests.BatchConsume
 {
     using System;
     using System.Threading.Tasks;
+    using FluentAssertions;
     using KafkaFlow.BatchConsume;
     using Microsoft.VisualStudio.TestTools.UnitTesting;
     using Moq;
@@ -9,87 +10,131 @@ namespace KafkaFlow.UnitTests.BatchConsume
     [TestClass]
     public class BatchConsumeMiddlewareTests
     {
-        private const int BatchSize = 10;
-        private readonly TimeSpan batchTimeout = TimeSpan.FromSeconds(3);
+        private const int BatchSize = 3;
 
-        private Mock<IWorkerBatchFactory> workerBatchFactoryMock;
+        private readonly TimeSpan batchTimeout = TimeSpan.FromMilliseconds(1000);
+        private readonly TimeSpan waitForTaskExecution = TimeSpan.FromMilliseconds(100);
+
         private Mock<ILogHandler> logHandlerMock;
+
+        private IMessageContext nextContext;
+        private int timesNextWasCalled;
 
         private BatchConsumeMiddleware target;
 
         [TestInitialize]
         public void Setup()
         {
-            this.workerBatchFactoryMock = new Mock<IWorkerBatchFactory>(MockBehavior.Strict);
+            this.nextContext = null;
+            this.timesNextWasCalled = 0;
+
             this.logHandlerMock = new Mock<ILogHandler>(MockBehavior.Strict);
 
             this.target = new BatchConsumeMiddleware(
                 BatchSize,
                 this.batchTimeout,
-                this.workerBatchFactoryMock.Object,
                 this.logHandlerMock.Object);
         }
 
         [TestMethod]
-        public async Task Invoke_DifferentWorkers_CallAddAsyncForEachOne()
+        public async Task AddAsync_LessThanBatchSize_CallNextOnTimeout()
         {
             // Arrange
-            var contextWorker1Mock = new Mock<IMessageContext>();
-            var contextWorker2Mock = new Mock<IMessageContext>();
+            var consumerContext = new Mock<IConsumerContext>();
+            var context = new Mock<IMessageContext>();
 
-            var consumerContextWorker1Mock = new Mock<IConsumerContext>();
-            var consumerContextWorker2Mock = new Mock<IConsumerContext>();
-
-            var worker1Batch = new Mock<IWorkerBatch>();
-            var worker2Batch = new Mock<IWorkerBatch>();
-
-            var nextMock = new Mock<MiddlewareDelegate>();
-
-            consumerContextWorker1Mock
-                .SetupGet(x => x.WorkerId)
-                .Returns(1);
-
-            contextWorker1Mock
-                .SetupGet(x => x.ConsumerContext)
-                .Returns(consumerContextWorker1Mock.Object);
-
-            consumerContextWorker1Mock.SetupSet(x => x.ShouldStoreOffset = false);
-
-            consumerContextWorker2Mock
-                .SetupGet(x => x.WorkerId)
-                .Returns(2);
-
-            contextWorker2Mock
-                .SetupGet(x => x.ConsumerContext)
-                .Returns(consumerContextWorker2Mock.Object);
-
-            consumerContextWorker2Mock.SetupSet(x => x.ShouldStoreOffset = false);
-
-            worker1Batch
-                .Setup(x => x.AddAsync(contextWorker1Mock.Object, nextMock.Object))
-                .Returns(Task.CompletedTask);
-
-            worker2Batch
-                .Setup(x => x.AddAsync(contextWorker2Mock.Object, nextMock.Object))
-                .Returns(Task.CompletedTask);
-
-            this.workerBatchFactoryMock
-                .SetupSequence(x => x.Create(BatchSize, this.batchTimeout, this.logHandlerMock.Object))
-                .Returns(worker1Batch.Object)
-                .Returns(worker2Batch.Object);
+            context
+                .Setup(x => x.ConsumerContext)
+                .Returns(consumerContext.Object);
 
             // Act
-            await this.target.Invoke(contextWorker1Mock.Object, nextMock.Object);
-            await this.target.Invoke(contextWorker2Mock.Object, nextMock.Object);
+            await this.target.Invoke(context.Object, this.NextCallback);
 
             // Assert
-            this.workerBatchFactoryMock.VerifyAll();
-            worker1Batch.VerifyAll();
-            worker2Batch.VerifyAll();
-            contextWorker1Mock.VerifyAll();
-            contextWorker2Mock.VerifyAll();
-            consumerContextWorker1Mock.VerifyAll();
-            consumerContextWorker2Mock.VerifyAll();
+            this.timesNextWasCalled.Should().Be(0);
+            await this.WaitBatchTimeoutAsync();
+            this.timesNextWasCalled.Should().Be(1);
+            consumerContext.Verify(x => x.StoreOffset(), Times.Once);
+        }
+
+        [TestMethod]
+        public async Task AddAsync_ExactlyBatchSize_CallNextInstantly()
+        {
+            // Arrange
+            var consumerContext = new Mock<IConsumerContext>();
+            var contextMock = new Mock<IMessageContext>();
+
+            contextMock
+                .Setup(x => x.ConsumerContext)
+                .Returns(consumerContext.Object);
+
+            // Act
+            for (var i = 0; i < BatchSize; i++)
+            {
+                await this.target.Invoke(contextMock.Object, this.NextCallback);
+            }
+
+            // Assert
+            this.timesNextWasCalled.Should().Be(1);
+            this.nextContext.GetMessagesBatch().Should().HaveCount(BatchSize);
+            consumerContext.Verify(x => x.StoreOffset(), Times.Exactly(BatchSize));
+        }
+
+        [TestMethod]
+        public async Task AddAsync_MoreThanBatchSize_CallNextInstantlyThenCallWhenTimeout()
+        {
+            // Arrange
+            var consumerContext = new Mock<IConsumerContext>();
+            var contextMock = new Mock<IMessageContext>();
+
+            contextMock
+                .Setup(x => x.ConsumerContext)
+                .Returns(consumerContext.Object);
+
+            // Act
+            for (var i = 0; i < BatchSize + 1; i++)
+            {
+                await this.target.Invoke(contextMock.Object, this.NextCallback);
+            }
+
+            // Assert
+            this.timesNextWasCalled.Should().Be(1);
+            this.nextContext.GetMessagesBatch().Should().HaveCount(BatchSize);
+            consumerContext.Verify(x => x.StoreOffset(), Times.Exactly(BatchSize));
+
+            await this.WaitBatchTimeoutAsync();
+            this.timesNextWasCalled.Should().Be(2);
+            consumerContext.Verify(x => x.StoreOffset(), Times.Exactly(BatchSize + 1));
+        }
+
+        [TestMethod]
+        public async Task AddAsync_NextThrowException_LogError()
+        {
+            // Arrange
+            var consumerContext = new Mock<IConsumerContext>();
+            var contextMock = new Mock<IMessageContext>();
+
+            contextMock
+                .Setup(x => x.ConsumerContext)
+                .Returns(consumerContext.Object);
+
+            var ex = new Exception();
+
+            // Act
+            await this.target.Invoke(contextMock.Object, _ => throw ex);
+
+            // Assert
+            await this.WaitBatchTimeoutAsync();
+            this.logHandlerMock.Verify(x => x.Error(It.IsAny<string>(), ex, It.IsAny<object>()), Times.Once);
+        }
+
+        private Task WaitBatchTimeoutAsync() => Task.Delay(this.batchTimeout + this.waitForTaskExecution);
+
+        private Task NextCallback(IMessageContext ctx)
+        {
+            this.nextContext = ctx;
+            this.timesNextWasCalled++;
+            return Task.CompletedTask;
         }
     }
 }
