+++ src/KafkaFlow.UnitTests/OffsetManagerTests.cs
@@ -25,20 +25,20 @@ namespace KafkaFlow.UnitTests
         }
 
         [TestMethod]
-        public void StoreOffset_WithInvalidTopicPartition_ShouldDoNothing()
+        public void MarkAsProcessed_WithInvalidTopicPartition_ShouldDoNothing()
         {
             // Arrange
             this.target.Enqueue(new TopicPartitionOffset(this.topicPartition, new Offset(1)));
 
             // Act
-            this.target.Commit(new TopicPartitionOffset(new TopicPartition("topic-B", new Partition(1)), new Offset(1)));
+            this.target.MarkAsProcessed(new TopicPartitionOffset(new TopicPartition("topic-B", new Partition(1)), new Offset(1)));
 
             // Assert
-            this.committerMock.Verify(c => c.Commit(It.IsAny<TopicPartitionOffset>()), Times.Never());
+            this.committerMock.Verify(c => c.MarkAsProcessed(It.IsAny<TopicPartitionOffset>()), Times.Never());
         }
 
         [TestMethod]
-        public void StoreOffset_WithGaps_ShouldStoreOffsetJustOnce()
+        public void MarkAsProcessed_WithGaps_ShouldStoreOffsetJustOnce()
         {
             // Arrange
             this.target.Enqueue(new TopicPartitionOffset(this.topicPartition, new Offset(1)));
@@ -46,14 +46,14 @@ namespace KafkaFlow.UnitTests
             this.target.Enqueue(new TopicPartitionOffset(this.topicPartition, new Offset(3)));
 
             // Act
-            this.target.Commit(new TopicPartitionOffset(this.topicPartition, new Offset(3)));
-            this.target.Commit(new TopicPartitionOffset(this.topicPartition, new Offset(2)));
-            this.target.Commit(new TopicPartitionOffset(this.topicPartition, new Offset(1)));
+            this.target.MarkAsProcessed(new TopicPartitionOffset(this.topicPartition, new Offset(3)));
+            this.target.MarkAsProcessed(new TopicPartitionOffset(this.topicPartition, new Offset(2)));
+            this.target.MarkAsProcessed(new TopicPartitionOffset(this.topicPartition, new Offset(1)));
 
             // Assert
             this.committerMock.Verify(
                 c =>
-                    c.Commit(
+                    c.MarkAsProcessed(
                         It.Is<TopicPartitionOffset>(
                             p =>
                                 p.Partition.Equals(this.topicPartition.Partition) &&
