+++ src/KafkaFlow/Consumers/IOffsetManager.cs
@@ -1,9 +1,12 @@
 namespace KafkaFlow.Consumers
 {
+    using System;
     using Confluent.Kafka;
 
     internal interface IOffsetManager
     {
-        void Commit(TopicPartitionOffset offset);
+        void MarkAsProcessed(TopicPartitionOffset offset);
+
+        void OnOffsetProcessed(TopicPartitionOffset offset, Action action);
     }
 }
